requirements.txt
Django==5.0.1
Pillow==10.1.0
qrcode==7.4.2 